+++ java-design-processor/src/test/resources/internal/nbbrd/design/TestDecoratorConstructorMissing.java
@@ -1,0 +1,10 @@
+package internal.nbbrd.design;
+
+import nbbrd.design.DecoratorPattern;
+
+import java.io.Reader;
+
+@DecoratorPattern
+public abstract class TestDecoratorConstructorMissing extends Reader {
+
+}