+++ java-design-processor/src/test/resources/internal/nbbrd/design/TestDecoratorConstructorNoArg.java
@@ -1,0 +1,11 @@
+package internal.nbbrd.design;
+
+import nbbrd.design.DecoratorPattern;
+
+import java.io.Reader;
+
+@DecoratorPattern
+public abstract class TestDecoratorConstructorNoArg extends Reader {
+
+    TestDecoratorConstructorNoArg() {}
+}